requirements.txt
matplotlib==3.7.1
pandas==2.0.1
scikit-learn==1.2.2
statsmodels==0.14.0


# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

# ML
ultralytics==8.3.0
numpy==1.26.4
Pillow==10.4.0
opencv-python-headless==4.10.0.84

# Testing
pytest==8.3.3
httpx==0.27.2          # for FastAPI TestClient

# Monitoring
prometheus-fastapi-instrumentator==7.0.0

# Frontend
streamlit==1.38.0
requests==2.32.3